fix(profile/widgets): refactor file to less than 700 lines by moving hooks to separate file
